refactor model parser from json; add writer for assertions on model
