+++ fluentgen/src/main/java/com/azure/autorest/fluent/mapper/ExampleParser.java
@@ -13,7 +13,6 @@ import com.azure.autorest.fluent.model.clientmodel.FluentResourceModel;
 import com.azure.autorest.fluent.model.clientmodel.FluentStatic;
 import com.azure.autorest.fluent.model.clientmodel.MethodParameter;
 import com.azure.autorest.model.clientmodel.ModelProperty;
-import com.azure.autorest.model.clientmodel.examplemodel.ClientModelNode;
 import com.azure.autorest.model.clientmodel.examplemodel.ExampleNode;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentClientMethodExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentCollectionMethodExample;
@@ -22,8 +21,6 @@ import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentResourceCr
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentResourceUpdateExample;
 import com.azure.autorest.model.clientmodel.examplemodel.ListNode;
 import com.azure.autorest.model.clientmodel.examplemodel.LiteralNode;
-import com.azure.autorest.model.clientmodel.examplemodel.MapNode;
-import com.azure.autorest.model.clientmodel.examplemodel.ObjectNode;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.ParameterExample;
 import com.azure.autorest.fluent.model.clientmodel.fluentmodel.create.DefinitionStage;
 import com.azure.autorest.fluent.model.clientmodel.fluentmodel.create.DefinitionStageBlank;
@@ -43,36 +40,34 @@ import com.azure.autorest.model.clientmodel.ClientMethod;
 import com.azure.autorest.model.clientmodel.ClientMethodParameter;
 import com.azure.autorest.model.clientmodel.ClientMethodType;
 import com.azure.autorest.model.clientmodel.ClientModel;
-import com.azure.autorest.model.clientmodel.ClientModelProperty;
 import com.azure.autorest.model.clientmodel.IType;
 import com.azure.autorest.model.clientmodel.ListType;
-import com.azure.autorest.model.clientmodel.MapType;
 import com.azure.autorest.model.clientmodel.MethodGroupClient;
 import com.azure.autorest.model.clientmodel.ProxyMethodExample;
 import com.azure.autorest.model.clientmodel.ProxyMethodParameter;
 import com.azure.autorest.util.CodeNamer;
-import com.azure.core.util.CoreUtils;
+import com.azure.autorest.util.ModelExampleUtil;
 import com.azure.core.util.serializer.CollectionFormat;
 import org.slf4j.Logger;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.function.Function;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-
 public class ExampleParser {
 
     private static final Logger logger = new PluginLogger(FluentGen.getPluginInstance(), ExampleParser.class);
+
+    static {
+        ModelExampleUtil.setGetClientModelFunction(FluentUtils::getClientModel);
+    }
 
     private final boolean aggregateExamples;
 
@@ -527,9 +522,9 @@ public class ExampleParser {
         } else {
             List<String> jsonPropertyNames = modelProperty.getSerializedNames();
 
-            Object childObjectValue = getChildObjectValue(jsonPropertyNames, parameterValue.getObjectValue());
+            Object childObjectValue = ModelExampleUtil.getChildObjectValue(jsonPropertyNames, parameterValue.getObjectValue());
             if (childObjectValue != null) {
-                node = parseNode(modelProperty.getClientType(), childObjectValue);
+                node = ModelExampleUtil.parseNode(modelProperty.getClientType(), childObjectValue);
             } else {
                 node = new LiteralNode(modelProperty.getClientType(), null);
             }
@@ -567,145 +562,13 @@ public class ExampleParser {
                     logger.error("Parameter style '{}' is not supported, fallback to CSV", collectionFormat);
             }
             for (String childObjectValue : elements) {
-                ExampleNode childNode = parseNode(elementType, childObjectValue);
+                ExampleNode childNode = ModelExampleUtil.parseNode(elementType, childObjectValue);
                 listNode.getChildNodes().add(childNode);
             }
             return listNode;
         } else {
-            return parseNode(type, objectValue);
+            return ModelExampleUtil.parseNode(type, objectValue);
         }
-    }
-
-    @SuppressWarnings("unchecked")
-    private static ExampleNode parseNode(IType type, Object objectValue) {
-        ExampleNode node;
-        if (type instanceof ListType) {
-            IType elementType = ((ListType) type).getElementType();
-            if (objectValue instanceof List) {
-                ListNode listNode = new ListNode(elementType, objectValue);
-                node = listNode;
-
-                List<Object> elements = (List<Object>) objectValue;
-                for (Object childObjectValue : elements) {
-                    ExampleNode childNode = parseNode(elementType, childObjectValue);
-                    node.getChildNodes().add(childNode);
-                }
-            } else {
-                logger.error("Example value is not List type: {}", objectValue);
-                node = new ListNode(elementType, null);
-            }
-        } else if (type instanceof MapType) {
-            IType elementType = ((MapType) type).getValueType();
-            if (objectValue instanceof Map) {
-                MapNode mapNode = new MapNode(elementType, objectValue);
-                node = mapNode;
-
-                Map<String, Object> dict = (Map<String, Object>) objectValue;
-                for (Map.Entry<String, Object> entry : dict.entrySet()) {
-                    ExampleNode childNode = parseNode(elementType, entry.getValue());
-                    node.getChildNodes().add(childNode);
-                    mapNode.getKeys().add(entry.getKey());
-                }
-            } else {
-                logger.error("Example value is not Map type: {}", objectValue);
-                node = new MapNode(elementType, null);
-            }
-        } else if (type == ClassType.Object) {
-            node = new ObjectNode(type, objectValue);
-        } else if (type instanceof ClassType && objectValue instanceof Map) {
-            ClientModel model = FluentUtils.getClientModel(((ClassType) type).getName());
-            if (model != null) {
-                if (model.getIsPolymorphic()) {
-                    // polymorphic, need to get the correct subclass from discriminator
-                    String serializedName = model.getPolymorphicDiscriminator();
-                    List<String> jsonPropertyNames = Collections.singletonList(serializedName);
-                    if (model.getNeedsFlatten()) {
-                        jsonPropertyNames = FluentUtils.splitFlattenedSerializedName(serializedName);
-                    }
-
-                    Object childObjectValue = getChildObjectValue(jsonPropertyNames, objectValue);
-                    if (childObjectValue instanceof String) {
-                        String discriminatorValue = (String) childObjectValue;
-                        ClientModel derivedModel = getDerivedModel(model, discriminatorValue);
-                        if (derivedModel != null) {
-                            // use the subclass
-                            type = derivedModel.getType();
-                            model = derivedModel;
-                        } else {
-                            logger.warn("Failed to find the subclass with discriminator value '{}'", discriminatorValue);
-                        }
-                    } else {
-                        logger.warn("Failed to find the sample value for discriminator property '{}'", serializedName);
-                    }
-                }
-
-                ClientModelNode clientModelNode = new ClientModelNode(type, objectValue).setClientModel(model);
-                node = clientModelNode;
-
-                List<ModelProperty> modelProperties = getWritablePropertiesIncludeSuperclass(model);
-                for (ModelProperty modelProperty : modelProperties) {
-                    List<String> jsonPropertyNames = modelProperty.getSerializedNames();
-
-                    Object childObjectValue = getChildObjectValue(jsonPropertyNames, objectValue);
-                    if (childObjectValue != null) {
-                        ExampleNode childNode = parseNode(modelProperty.getClientType(), childObjectValue);
-                        node.getChildNodes().add(childNode);
-                        clientModelNode.getClientModelProperties().put(childNode, modelProperty);
-                    }
-                }
-
-                // additional properties
-                ModelProperty additionalPropertiesProperty = getAdditionalPropertiesProperty(model);
-                if (additionalPropertiesProperty != null) {
-                    // properties already defined in model
-                    Set<String> propertySerializedNames = modelProperties.stream()
-                            .map(p -> p.getSerializedNames().iterator().next())
-                            .collect(Collectors.toSet());
-                    // the remaining properties in json
-                    Map<String, Object> remainingValues = ((Map<String, Object>) objectValue).entrySet().stream()
-                            .filter(e -> !propertySerializedNames.contains(e.getKey()))
-                            .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-
-                    ExampleNode childNode = parseNode(additionalPropertiesProperty.getClientType(), remainingValues);
-                    node.getChildNodes().add(childNode);
-                    clientModelNode.getClientModelProperties().put(childNode, additionalPropertiesProperty);
-                }
-            } else {
-                throw new IllegalStateException("Model type not found for type " + type + " and value " + objectValue);
-            }
-        } else if (objectValue == null) {
-            node = null;
-        } else {
-            LiteralNode literalNode = new LiteralNode(type, objectValue);
-            node = literalNode;
-
-            literalNode.setLiteralsValue(objectValue.toString());
-        }
-        return node;
-    }
-
-    private static Object getChildObjectValue(List<String> jsonPropertyNames, Object objectValue) {
-        boolean found = true;
-        Object childObjectValue = objectValue;
-        // walk the sequence of serialized names
-        for (String name : jsonPropertyNames) {
-            if (name.isEmpty()) {
-                found = false;
-                break;
-            }
-
-            if (childObjectValue instanceof Map) {
-                childObjectValue = ((Map<String, Object>) childObjectValue).get(name);
-                if (childObjectValue == null) {
-                    found = false;
-                    break;
-                }
-            } else {
-                found = false;
-                break;
-            }
-        }
-        return found ? childObjectValue : null;
     }
 
     private static List<MethodParameter> getParameters(ClientMethod clientMethod) {
@@ -716,83 +579,6 @@ public class ExampleParser {
                 .filter(p -> !p.getIsConstant() && !p.getFromClient())
                 .map(p -> new MethodParameter(proxyMethodParameterByClientParameterName.get(p.getName()), p))
                 .collect(Collectors.toList());
-    }
-
-    private static ModelProperty getAdditionalPropertiesProperty(ClientModel model) {
-        ModelProperty modelProperty = null;
-        ClientModelProperty property = model.getProperties().stream()
-                .filter(ClientModelProperty::isAdditionalProperties)
-                .findFirst().orElse(null);
-        if (property != null && property.getClientType() instanceof MapType) {
-            modelProperty = ModelProperty.ofClientModelProperty(property);
-        }
-        return modelProperty;
-    }
-
-    private static List<ModelProperty> getWritablePropertiesIncludeSuperclass(ClientModel model) {
-        Map<String, ModelProperty> propertiesMap = new LinkedHashMap<>();
-        List<ModelProperty> properties = new ArrayList<>();
-
-        List<ClientModel> parentModels = new ArrayList<>();
-        String parentModelName = model.getParentModelName();
-        while (!CoreUtils.isNullOrEmpty(parentModelName)) {
-            ClientModel parentModel = FluentUtils.getClientModel(parentModelName);
-            if (parentModel != null) {
-                parentModels.add(parentModel);
-            }
-            parentModelName = parentModel == null ? null : parentModel.getParentModelName();
-        }
-
-        List<List<ModelProperty>> propertiesFromTypeAndParents = new ArrayList<>();
-        propertiesFromTypeAndParents.add(new ArrayList<>());
-        model.getAccessibleProperties().forEach(p -> {
-            ModelProperty modelProperty = ModelProperty.ofClientModelProperty(p);
-            if (propertiesMap.putIfAbsent(modelProperty.getName(), modelProperty) == null) {
-                propertiesFromTypeAndParents.get(propertiesFromTypeAndParents.size() - 1).add(modelProperty);
-            }
-        });
-
-        for (ClientModel parent : parentModels) {
-            propertiesFromTypeAndParents.add(new ArrayList<>());
-
-            parent.getAccessibleProperties().forEach(p -> {
-                ModelProperty modelProperty = ModelProperty.ofClientModelProperty(p);
-                if (propertiesMap.putIfAbsent(modelProperty.getName(), modelProperty) == null) {
-                    propertiesFromTypeAndParents.get(propertiesFromTypeAndParents.size() - 1).add(modelProperty);
-                }
-            });
-        }
-
-        Collections.reverse(propertiesFromTypeAndParents);
-        for (List<ModelProperty> properties1 : propertiesFromTypeAndParents) {
-            properties.addAll(properties1);
-        }
-
-        return properties.stream()
-                .filter(p -> !p.isReadOnly() && !p.isConstant())
-                .collect(Collectors.toList());
-    }
-
-    private static ClientModel getDerivedModel(ClientModel model, String discriminatorValue) {
-        // depth first search
-
-        if (model.getDerivedModels() != null) {
-            for (ClientModel childModel : model.getDerivedModels()) {
-                if (discriminatorValue.equalsIgnoreCase(childModel.getSerializedName())) {
-                    // found
-                    return childModel;
-                } else if (childModel.getDerivedModels() != null) {
-                    // recursive
-                    ClientModel childModel2 = getDerivedModel(childModel, discriminatorValue);
-                    if (childModel2 != null) {
-                        return childModel2;
-                    }
-                }
-            }
-        }
-
-        // not found
-        return null;
     }
 
     private static boolean requiresExample(ClientMethod clientMethod) {
